% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_presenter_name.R
\name{add_presenter_name}
\alias{add_presenter_name}
\title{Add presenter name to qr_code}
\usage{
add_presenter_name(presenter_name, png, poster_id)
}
\arguments{
\item{presenter_name}{The name of the presenter you'd like to add to the qr code}

\item{png}{a file path to a png you'd like to add a label to}

\item{poster_id}{The associated id for the poster}
}
\value{
A png with a name label on the top
}
\description{
This returns a random jumble of letters and numbers that are used as poster ids
}
