% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_poster_data_file_path}
\alias{example_poster_data_file_path}
\title{Poster Data Example File
Get file path to an example poster data file}
\usage{
example_poster_data_file_path()
}
\value{
Returns the file path to folder where the example data is stored
}
\description{
Poster Data Example File
Get file path to an example poster data file
}
