% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_forms.R
\name{make_poll}
\alias{make_poll}
\title{Setup posterpoller form}
\usage{
make_poll(form_id, new_name = NULL, quiet = FALSE)
}
\arguments{
\item{form_id}{The form_id that is desired to be copied.}

\item{new_name}{What should the new file name for the copied file be?}

\item{quiet}{TRUE or FALSE whether messages should be printed out.}
}
\description{
Setup posterpoller form
}
\examples{
\dontrun{
#'
# Make the form
form_info <- make_poll(form_id = "https://docs.google.com/forms/d/someformidhere/edit",
                       new_name = "copied form")
}
}
