% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_poster_ids.R
\name{make_random_ids}
\alias{make_random_ids}
\title{Make a random id}
\usage{
make_random_ids(n = 1, seed = 1234)
}
\arguments{
\item{n}{How many ids do you want returned?}

\item{seed}{Set a seed so that the IDs will be the same if you rerun this}
}
\value{
A vector of length n with that many random ids
}
\description{
This returns a random jumble of letters and numbers that are used as poster ids
}
